refactor(react/runtime): add new lifecycle/patch folder
