+++ packages/react/runtime/__test__/snapshot/ref.test.jsx
@@ -5,13 +5,14 @@
 */
 import { render } from 'preact';
 import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from 'vitest';
+
 import { Component, createRef, useState } from '../../src/index';
+import { injectUpdatePatch, replaceCommitHook } from '../../src/lifecycle/patch/patchUpdate';
+import { __pendingListUpdates } from '../../src/list';
 import { __root } from '../../src/root';
 import { setupPage } from '../../src/snapshot';
 import { globalEnvManager } from '../utils/envManager';
 import { elementTree, waitSchedule } from '../utils/nativeMethod';
-import { injectUpdatePatch, replaceCommitHook } from '../../src/lifecycle/patchUpdate';
-import { __pendingListUpdates } from '../../src/list';
 
 beforeAll(() => {
   setupPage(__CreatePage('0', 0));
