+++ packages/react/runtime/__test__/snapshot/workletEvent.test.jsx
@@ -5,12 +5,13 @@
 */
 import { render } from 'preact';
 import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from 'vitest';
+
+import { Component, useState } from '../../src/index';
+import { injectUpdatePatch, replaceCommitHook } from '../../src/lifecycle/patch/patchUpdate';
+import { __root } from '../../src/root';
 import { setupPage } from '../../src/snapshot';
 import { globalEnvManager } from '../utils/envManager';
 import { elementTree, waitSchedule } from '../utils/nativeMethod';
-import { __root } from '../../src/root';
-import { injectUpdatePatch, replaceCommitHook } from '../../src/lifecycle/patchUpdate';
-import { Component, useState } from '../../src/index';
 
 beforeAll(() => {
   setupPage(__CreatePage('0', 0));
