+++ packages/react/runtime/__test__/snapshot/workletRef.test.jsx
@@ -5,13 +5,14 @@
 */
 import { render } from 'preact';
 import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from 'vitest';
+
+import { createCompBG1, createCompBGList, createCompBGSpread } from './workletRefBG';
+import { createCompMT1, createCompMTList, createCompMTSpread } from './workletRefMT';
+import { injectUpdatePatch, replaceCommitHook } from '../../src/lifecycle/patch/patchUpdate';
+import { __root } from '../../src/root';
 import { setupPage } from '../../src/snapshot';
 import { globalEnvManager } from '../utils/envManager';
 import { elementTree } from '../utils/nativeMethod';
-import { __root } from '../../src/root';
-import { injectUpdatePatch, replaceCommitHook } from '../../src/lifecycle/patchUpdate';
-import { createCompBG1, createCompBGList, createCompBGSpread } from './workletRefBG';
-import { createCompMT1, createCompMTList, createCompMTSpread } from './workletRefMT';
 
 beforeAll(() => {
   setupPage(__CreatePage('0', 0));
