+++ packages/react/runtime/src/lifecycle/patch/snapshotPatch.ts
