+++ packages/react/runtime/src/worklet/hmr.ts
@@ -1,7 +1,7 @@
 // Copyright 2024 The Lynx Authors. All rights reserved.
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
-import { SnapshotOperation, __globalSnapshotPatch } from '../snapshotPatch.js';
+import { SnapshotOperation, __globalSnapshotPatch } from '../lifecycle/patch/snapshotPatch.js';
 
 const workletHashSet: Set<string> = /* @__PURE__ */ new Set();
